+++ tests/methods/__init__.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+
+__all__ = [
+    'PlanToConfigurationTest',
+    'PlanToEndEffectorPoseTest',
+    'RetimeTrajectoryTest',
+    'ShortcutPathTest',
+    'SmoothTrajectoryTest',
+]
+
+from PlanToConfiguration import PlanToConfigurationTest
+from PlanToEndEffectorPose import PlanToEndEffectorPoseTest
+from RetimeTrajectory import RetimeTrajectoryTest
+from ShortcutPath import ShortcutPathTest
+from SmoothTrajectory import SmoothTrajectoryTest
